+++ tools/devserver/schemas/user.js
@@ -19,14 +19,19 @@ module.exports = function(server){
 
 		username: joi.string().alphanum().min(3).max(36).required(),
 		password: joi.string().min(6).max(36), //will be hashed by salt later
-		userspace: joi.string().valid('user', 'admin').default('user'),
-		email: joi.string().email() //for account activation, retrieval.
-		//profile: -- [optional child schema]
+		email: joi.string().email(), //for account activation, retrieval.
+		suspended: joi.boolean().default(false),
+		//profile: -- additional info about this user [optional child schema]
 		//------------------------------------------------------------
 		//salt: -- auto-generated in routers.user to hash the password
 		//------------------------------------------------------------
-		//space: record space
-		//last_login: timestamp
+		userspace: joi.string().valid('user', 'admin').default('user'),
+		//space: -- record space, imposed upon record saving
+		//------------------------------------------------------------
+		//roles: -- role names, all roles (api-tokens) combined together = user api permission
+		//------------------------------------------------------------
+		
+		//last_logon: timestamp
 		//created_at: timestamp
 		//updated_at: timestamp